Compute type_equal() without reference to backend()
Argument for correctness: type_equal previous compared if backends
are equal.  Backend is computed by translation from dispatch key.
I verified that computeDispatchKey never computed a weird
dispatch key (e.g., AutogradXLA), so that dispatchKeyToBackend
was effectively injective.  Then it is always valid to compare
the arguments of an injective function for equality, rather than
the output of the injective function.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 8714dd28c2d2d749842a6fa0845c86697d2320a0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53823